+++ pytsite/util/__init__.py
@@ -2,7 +2,9 @@
 """
 # Public API
 from werkzeug.urls import url_quote, url_unquote
-from ._functions import *
+from ._api import strip_html_tags, cleanup_dict, cleanup_list, dict_merge, escape_html, get_class, html_attrs_str, \
+    md5_hex_digest, mk_tmp_file, nav_link, random_password, random_str, rfc822_datetime, transform_str_1, \
+    transform_str_2, transliterate, trim_str, w3c_datetime, weight_sort
 
 __author__ = 'Alexander Shepetko'
 __email__ = 'a@shepetko.com'